+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 0.2 (Unreleased)
 
+* Added `.results` shortcut for `new(*arg).results`
+
 * Fix wrong limit and offset in default paging
 
 ## Version 0.1